requirement.txt
langchain_groq
langchain
python-dotenv
langchain-community
langchain_core
streamlit
bs4
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
faiss-cpu
langchain_chroma
duckdb
pandas
openai
arxiv
wikipedia
SQLAlchemy
validators==0.28.1
youtube_transcript_api
numexpr